requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
anyio==3.4.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
async-generator==1.10
async-timeout==4.0.1
atomicwrites==1.1.5
attrs==19.3.0
Automat==20.2.0
autopage==0.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
build==0.7.0
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.4
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.3
colorlog==6.6.0
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-js-asset==2.0.0
djangorestframework==3.13.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.18.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.2.0
fsspec==2021.11.0
functions==0.7.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
h11==0.13.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==21.3.0
install==1.3.5
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.1.1
jdcal==1.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
Js2Py==0.71
json5==0.9.6
jsonschema==4.2.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightgbm==3.3.1
lockfile==0.12.2
logger==1.4
loguru==0.5.3
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
paramiko==2.6.0
parsel==1.6.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.2
pbr==5.8.0
pep517==0.12.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.4.0
pipwin==0.5.2
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==0.13.0
prettytable==2.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
Protego==0.2.1
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.8.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
pyDeprecate==0.3.1
PyDispatcher==2.0.5
Pygments==2.10.0
PyGObject==3.36.0
pyjsparser==2.7.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyorbital==1.7.1
pyparsing==2.4.6
pyperclip==1.8.2
PyPrind==2.11.3
pyRFC3339==1.1
pyrsistent==0.18.0
pySmartDL==1.3.4
PySocks==1.7.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
queuelib==1.6.2
readme-renderer==30.0
regex==2022.3.15
reportlab==3.5.34
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.7.2
Send2Trash==1.8.0
sentry-sdk==1.5.7
service-identity==21.1.0
setproctitle==1.2.2
shortuuid==1.0.8
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
SQLAlchemy==1.4.27
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2